requirements.txt
pandas==1.4.2
numpy==1.22.4
python-dotenv==0.20.0
selenium==4.1.3